Add client working area, etc
Each client will have its own working area to store things like client-duid,
pid file, log, and DBs.

There is a change in the server side to make sure that answers and requests
are matched not only on the transaction id basis, but also on the Peer
address. It was found that the same transaction id may be used by multiple
clients at the same time.